Move worker primary net to network crate
