+++ worker/src/primary_connector.rs
@@ -3,17 +3,13 @@
 // SPDX-License-Identifier: Apache-2.0
 use config::Committee;
 use crypto::traits::VerifyingKey;
-use futures::{stream::FuturesUnordered, FutureExt, StreamExt};
-use multiaddr::Multiaddr;
-use network::{BoundedExecutor, CancelHandler, RetryConfig, MAX_TASK_CONCURRENCY};
-use primary::WorkerPrimaryMessage;
+use futures::{stream::FuturesUnordered, StreamExt};
+use network::WorkerToPrimaryNetwork;
 use tokio::{
-    runtime::Handle,
     sync::{mpsc::Receiver, watch},
     task::JoinHandle,
 };
-use tonic::transport::Channel;
-use types::{BincodeEncodedPayload, ReconfigureNotification, WorkerToPrimaryClient};
+use types::{ReconfigureNotification, WorkerPrimaryMessage};
 
 /// The maximum number of digests kept in memory waiting to be sent to the primary.
 pub const MAX_PENDING_DIGESTS: usize = 10_000;
@@ -86,69 +82,5 @@ impl<PublicKey: VerifyingKey> PrimaryConnector<PublicKey> {
                 Some(()) = futures.next() => ()
             }
         }
-    }
-}
-
-pub struct WorkerToPrimaryNetwork {
-    address: Option<Multiaddr>,
-    client: Option<WorkerToPrimaryClient<Channel>>,
-    config: mysten_network::config::Config,
-    retry_config: RetryConfig,
-    executor: BoundedExecutor,
-}
-
-impl Default for WorkerToPrimaryNetwork {
-    fn default() -> Self {
-        let retry_config = RetryConfig {
-            // Retry for forever
-            retrying_max_elapsed_time: None,
-            ..Default::default()
-        };
-
-        Self {
-            address: None,
-            client: Default::default(),
-            config: Default::default(),
-            retry_config,
-            executor: BoundedExecutor::new(MAX_TASK_CONCURRENCY, Handle::current()),
-        }
-    }
-}
-
-impl WorkerToPrimaryNetwork {
-    pub async fn send<PublicKey: VerifyingKey>(
-        &mut self,
-        address: Multiaddr,
-        message: &WorkerPrimaryMessage<PublicKey>,
-    ) -> CancelHandler<()> {
-        let new_client = match &self.address {
-            None => true,
-            Some(x) if x != &address => true,
-            _ => false,
-        };
-        if new_client {
-            let channel = self.config.connect_lazy(&address).unwrap();
-            self.client = Some(WorkerToPrimaryClient::new(channel));
-            self.address = Some(address);
-        }
-
-        let message =
-            BincodeEncodedPayload::try_from(message).expect("Failed to serialize payload");
-        let client = self.client.as_mut().unwrap().clone();
-        let handle = self
-            .executor
-            .spawn(
-                self.retry_config
-                    .retry(move || {
-                        let mut client = client.clone();
-                        let message = message.clone();
-                        async move { client.send_message(message).await.map_err(Into::into) }
-                    })
-                    .map(|response| {
-                        response.expect("we retry forever so this shouldn't fail");
-                    }),
-            )
-            .await;
-        CancelHandler(handle)
     }
 }
